fulfillment/shrivu/requirements.txt
python-dateutil==2.6.0
requests